+++ prometheus/autometrics/prometheus.go
@@ -1,10 +1,17 @@
 package autometrics // import "github.com/autometrics-dev/autometrics-go/prometheus/autometrics"
 
 import (
+	"context"
+	"errors"
+	"fmt"
+	"log"
 	"os"
+	"sync"
 
 	"github.com/autometrics-dev/autometrics-go/pkg/autometrics"
 	"github.com/prometheus/client_golang/prometheus"
+	"github.com/prometheus/client_golang/prometheus/push"
+	"github.com/prometheus/common/expfmt"
 )
 
 var (
@@ -13,6 +20,10 @@ var (
 	functionCallsConcurrent *prometheus.GaugeVec
 	buildInfo               *prometheus.GaugeVec
 	DefBuckets              = autometrics.DefBuckets
+
+	amCtx      context.Context
+	pusher     *push.Pusher
+	pusherLock sync.Mutex
 )
 
 const (
@@ -68,6 +79,13 @@ const (
 	// ServiceNameLabel is the prometheus label that describes the name of the service being monitored
 	ServiceNameLabel = "service_name"
 
+	// ClearModeLabel is the label used by Prometheus Gravel Gateway to deal with aggregation.
+	ClearModeLabel = "clearmode"
+
+	ClearModeFamily    = "family"
+	ClearModeAggregate = "aggregate"
+	ClearModeReplace   = "replace"
+
 	traceIdExemplar      = "trace_id"
 	spanIdExemplar       = "span_id"
 	parentSpanIdExemplar = "parent_id"
@@ -79,19 +97,60 @@ const (
 // the current (prometheus) package imported at the call site.
 type BuildInfo = autometrics.BuildInfo
 
+// PushConfiguration holds meta information about the push-to-collector configuration of the instrumented code.
+//
+// This is a reexport of the autometrics type to allow [Init] to work with only
+// the current (prometheus) package imported at the call site.
+//
+// For the CollectorURL part, just as the prometheus library [push] configuration,
+// "You can use just host:port or ip:port as url, in which case “http://” is
+// added automatically. Alternatively, include the schema in the URL. However,
+// do not include the “/metrics/jobs/…” part."
+//
+// [push]: https://pkg.go.dev/github.com/prometheus/client_golang/prometheus/push#New
+type PushConfiguration = autometrics.PushConfiguration
+
 // Init sets up the metrics required for autometrics' decorated functions and registers
 // them to the argument registry.
 //
 // If the passed registry is nil, all the metrics are registered to the
 // default global registry.
 //
+// After initialization, use the returned [context.CancelCauseFunc] to flush the last
+// results and turn off metric collection for the remainder of the program's lifetime.
+// It is a good candidate to be deferred in the usual case.
+//
 // Make sure that all the latency targets you want to use for SLOs are
 // present in the histogramBuckets array, otherwise the alerts will fail
 // to work (they will never trigger.)
-func Init(reg *prometheus.Registry, histogramBuckets []float64, buildInformation BuildInfo) error {
+func Init(reg *prometheus.Registry, histogramBuckets []float64, buildInformation BuildInfo, pushConfiguration *PushConfiguration) (context.CancelCauseFunc, error) {
+	newCtx, cancelFunc := context.WithCancelCause(context.Background())
+	amCtx = newCtx
+
 	autometrics.SetCommit(buildInformation.Commit)
 	autometrics.SetVersion(buildInformation.Version)
 	autometrics.SetBranch(buildInformation.Branch)
+
+	pusher = nil
+	if pushConfiguration != nil {
+		log.Printf("autometrics: Init: detected push configuration to %s", pushConfiguration.CollectorURL)
+
+		if pushConfiguration.CollectorURL == "" {
+			return nil, errors.New("invalid PushConfiguration: the CollectorURL must be set.")
+		}
+		autometrics.SetPushJobURL(pushConfiguration.CollectorURL)
+
+		if pushConfiguration.JobName == "" {
+			autometrics.SetPushJobName(autometrics.DefaultJobName())
+		} else {
+			autometrics.SetPushJobName(pushConfiguration.JobName)
+		}
+
+		pusher = push.
+			New(autometrics.GetPushJobURL(), autometrics.GetPushJobName()).
+			Format(expfmt.FmtText)
+
+	}
 
 	if serviceName, ok := os.LookupEnv(autometrics.AutometricsServiceNameEnv); ok {
 		autometrics.SetService(serviceName)
@@ -103,20 +162,20 @@ func Init(reg *prometheus.Registry, histogramBuckets []float64, buildInformation
 
 	functionCallsCount = prometheus.NewCounterVec(prometheus.CounterOpts{
 		Name: FunctionCallsCountName,
-	}, []string{FunctionLabel, ModuleLabel, CallerFunctionLabel, CallerModuleLabel, ResultLabel, TargetSuccessRateLabel, SloNameLabel, CommitLabel, VersionLabel, BranchLabel, ServiceNameLabel})
+	}, []string{FunctionLabel, ModuleLabel, CallerFunctionLabel, CallerModuleLabel, ResultLabel, TargetSuccessRateLabel, SloNameLabel, CommitLabel, VersionLabel, BranchLabel, ServiceNameLabel, ClearModeLabel})
 
 	functionCallsDuration = prometheus.NewHistogramVec(prometheus.HistogramOpts{
 		Name:    FunctionCallsDurationName,
 		Buckets: histogramBuckets,
-	}, []string{FunctionLabel, ModuleLabel, CallerFunctionLabel, CallerModuleLabel, TargetLatencyLabel, TargetSuccessRateLabel, SloNameLabel, CommitLabel, VersionLabel, BranchLabel, ServiceNameLabel})
+	}, []string{FunctionLabel, ModuleLabel, CallerFunctionLabel, CallerModuleLabel, TargetLatencyLabel, TargetSuccessRateLabel, SloNameLabel, CommitLabel, VersionLabel, BranchLabel, ServiceNameLabel, ClearModeLabel})
 
 	functionCallsConcurrent = prometheus.NewGaugeVec(prometheus.GaugeOpts{
 		Name: FunctionCallsConcurrentName,
-	}, []string{FunctionLabel, ModuleLabel, CallerFunctionLabel, CallerModuleLabel, CommitLabel, VersionLabel, BranchLabel, ServiceNameLabel})
+	}, []string{FunctionLabel, ModuleLabel, CallerFunctionLabel, CallerModuleLabel, CommitLabel, VersionLabel, BranchLabel, ServiceNameLabel, ClearModeLabel})
 
 	buildInfo = prometheus.NewGaugeVec(prometheus.GaugeOpts{
 		Name: BuildInfoName,
-	}, []string{CommitLabel, VersionLabel, BranchLabel, ServiceNameLabel})
+	}, []string{CommitLabel, VersionLabel, BranchLabel, ServiceNameLabel, ClearModeLabel})
 
 	if reg != nil {
 		reg.MustRegister(functionCallsCount)
@@ -135,7 +194,50 @@ func Init(reg *prometheus.Registry, histogramBuckets []float64, buildInformation
 		VersionLabel:     buildInformation.Version,
 		BranchLabel:      buildInformation.Branch,
 		ServiceNameLabel: autometrics.GetService(),
+		ClearModeLabel:   ClearModeFamily,
 	}).Set(1)
+
+	if pusher != nil {
+		pusherLock.Lock()
+		defer pusherLock.Unlock()
+
+		if err := pusher.
+			Collector(buildInfo).
+			AddContext(amCtx); err != nil {
+			return nil, fmt.Errorf("pushing metrics to gateway for initialization: %w", err)
+		}
+	}
+
+	return cancelFunc, nil
+}
+
+// ForceFlush forces a flush of the metrics, in the case autometrics is pushing metrics to a Prometheus Push Gateway.
+//
+// This function is a no-op if no push configuration has been setup in [Init], but will return an error if
+// autometrics is not active (because this function is called before [Init] or after its shutdown function
+// has been called).
+func ForceFlush() error {
+	if amCtx.Err() != nil {
+		return fmt.Errorf("autometrics is not currently active: %w", amCtx.Err())
+	}
+
+	if pusher != nil {
+		ctx, cancel := context.WithCancel(amCtx)
+		defer cancel()
+		if pusherLock.TryLock() {
+			defer pusherLock.Unlock()
+			localPusher := push.
+				New(autometrics.GetPushJobURL(), autometrics.GetPushJobName()).
+				Format(expfmt.FmtText).
+				Collector(functionCallsCount).
+				Collector(functionCallsDuration).
+				Collector(functionCallsConcurrent)
+			if err := localPusher.
+				AddContext(ctx); err != nil {
+				return fmt.Errorf("failed to push metrics to gateway: %w", err)
+			}
+		}
+	}
 
 	return nil
 }